Implemented More functions
Made it so that the Add Hours to An Existing Job actually worked, along with the clock in and out function. Also made it so that the time for clock in must be before clock out time. Also added the columns of jobs at the bottom, which shows the last 6 shifts worked. Will add undo/remove shifts later for when you make a mistake